third_party/blink/web_tests/platform/generic/printing/css2.1/page-break-after-004-expected.txt

This test must produce two pages of output. This sentence must be on the first page.
This sentence must be at the top of the second page.
Lorem ipsum dolor sit amet, consectetuer adipiscing elit.

This must be the last sentence on the second and final page.


PASS: page number of "test1" is 1
PASS: page number of "test2" is 1
PASS: number of pages is 2
All tests passed
